// plugins/plugin-harbor/src/utils/interfaces.ts
export interface IOptions {
  query: string;
  page: number;
  perPage: number;
}
// ...
export interface IProjectsData {
  projects: IProject[];
  total: number;
}

// IProject is the interface of a single project returned by the Harbor API.
export interface IProject {
  name: string;
  deleted: boolean;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  repo_count: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  chart_count: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  project_id: number;
  // eslint-disable-next-line @typescript-eslint/naming-convention
  registry_id: number;
  metadata: {
    // eslint-disable-next-line @typescript-eslint/naming-convention
    enable_content_trust: string;
    // eslint-disable-next-line @typescript-eslint/naming-convention
    auto_scan: string;
    severity: string;
    public: string;
    // eslint-disable-next-line @typescript-eslint/naming-convention
    reuse_sys_cve_allowlist: string;
    // eslint-disable-next-line @typescript-eslint/naming-convention
    prevent_vul: string;
    // eslint-disable-next-line @typescript-eslint/naming-convention
    retention_id: string;
  };
}